Add post method for new experiments, simple test for controller, quick fix for JSON infinite recursion problem, a few corrections in get mappings.

+++ server/src/main/java/com/moea/repository/ExperimentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExperimentRepository extends JpaRepository<Experiment, Integer> {
+public interface ExperimentRepository extends JpaRepository<Experiment, Long> {
 }
